remove new test file, fix test cases for named parameters
