+++ test/py/test.py
@@ -1,16 +1,30 @@
-from time import sleep
+import argparse
 import os
 import threading
 from multiprocessing import get_context
+from time import sleep
+
+import h5py
 import numpy as np
 import PIL.Image as im
-
 from dlio_profiler.logger import dlio_logger, fn_interceptor
 
 cwd = os.getcwd()
 log_file = os.getenv("LOG_FILE", f"{cwd}/test_py-app.pwf")
 log_inst = dlio_logger.initialize_log(logfile=None, data_dir=None, process_id=-1)
 dlio_log = fn_interceptor("COMPUTE")
+
+parser = argparse.ArgumentParser(
+    prog='DLIO testing',
+    description='What the program does',
+    epilog='Text at the bottom of help')
+parser.add_argument("--data_dir", default="./data", type=str, help="The directory to save and load data")
+parser.add_argument("--format", default="npz", type=str, help="format of the file")
+parser.add_argument("--num_files", default=1, type=int, help="Number of files")
+parser.add_argument("--niter", default=1, type=int, help="Number of iterations for the experiment")
+parser.add_argument("--record_size", default=1048576, type=int, help="size of the record to be written to the file")
+args = parser.parse_args()
+os.makedirs(f"{args.data_dir}/{args.format}", exist_ok=True)
 
 
 @dlio_log.log
@@ -58,20 +72,80 @@ def jpeg_calls(index):
     records = np.random.randint(255, size=(1024, 1024), dtype=np.uint8)
     img = im.fromarray(records)
     out_path_spec = f"{cwd}/data/test.jpeg"
-    img.save(out_path_spec, format='JPEG', bits=8)
+    img.save(out_path_spec, format="JPEG", bits=8)
     with open(out_path_spec, "rb") as f:
         image = im.open(f)
         out_records = np.asarray(image)
     # image = im.open(out_path_spec)
 
 
+@dlio_log.log
+def data_gen(num_files, data_dir, format, data):
+    for i in dlio_log.iter(range(num_files)):
+        i = IOWriter(
+            filename=f"{data_dir}/{format}/{i}-of-{num_files}.{format}",
+            format=format,
+            data=data,
+        )
+
+
+@dlio_log.log
+def read_data(num_files, data_dir, format):
+    for i in dlio_log.iter(range(num_files)):
+        io = IOReader(
+            filename=f"{data_dir}/{format}/{i}-of-{num_files}.{format}",
+            format=format,
+        )
+        d = io.get()
+
+
+class IOHandler:
+    def __init__(self, filename, format):
+        self.format = format
+        self.filename = filename
+
+
+class IOReader(IOHandler):
+    @dlio_log.log_init
+    def __init__(self, filename, format):
+        super().__init__(filename=filename, format=format)
+        if self.format == "jpeg" or self.format == "png":
+            self.data = np.asarray(im.open(filename))
+        if self.format == "npz":
+            self.data = np.load(filename)
+        if self.format == "hdf5":
+            fd = h5py.File(filename, "r")
+            self.data = fd["x"][:]  # type: ignore
+            fd.close()
+
+    def get(self):
+        return self.data
+
+
+class IOWriter(IOHandler):
+    @dlio_log.log_init
+    def __init__(self, filename, format, data):
+        super().__init__(filename=filename, format=format)
+        if self.format == "jpeg" or self.format == "png":
+            image = im.fromarray(data)
+            # im.show()
+            image.save(filename)
+        if self.format == "npz":
+            with open(filename, "wb") as f:
+                np.save(f, data)
+        if self.format == "hdf5":
+            fd = h5py.File(filename, "w")
+            fd.create_dataset("x", data=data)
+            fd.close()
+
+
 def init():
     """This function is called when new processes start."""
-    print(f'Initializing process {os.getpid()}')
+    print(f"Initializing process {os.getpid()}")
 
 
 def main():
-    posix_calls((20,False))
+    posix_calls((20, False))
     t1 = threading.Thread(target=posix_calls, args=((10, False),))
     custom_events()
     t2 = threading.Thread(target=npz_calls, args=(1,))
@@ -88,13 +162,19 @@ def main():
     t3.join()
     t4.join()
     index = 4
-    with get_context('fork').Pool(1, initializer=init) as pool:
+    with get_context("fork").Pool(1, initializer=init) as pool:
         pool.map(posix_calls, ((index, False),))
     index = index + 1
 
-    with get_context('spawn').Pool(1, initializer=init) as pool:
+    with get_context("spawn").Pool(1, initializer=init) as pool:
         pool.map(posix_calls, ((index, True),))
     index = index + 1
+
+    # testing named parameters
+    data = np.ones((args.record_size, 1), dtype=np.uint8)
+    data_gen(num_files=args.num_files, data_dir=args.data_dir, format=args.format, data=data)
+    for n in range(args.niter):
+        read_data(num_files=args.num_files, data_dir=args.data_dir, format=args.format)
 
     log_inst.finalize()
